REFACTOR: Split Settings.ts to reduce number of imports
